+++ src/app/enums/power-mode.enum.ts
@@ -1,0 +1,6 @@
+export enum PowerMode {
+  Never = 'NEVER',
+  Sleep = 'SLEEP',
+  Standby = 'STANDBY',
+  Idle = 'IDLE',
+}
